+++ src/i18n/bg/nav.ts
@@ -1,4 +1,4 @@
-import { NavDictionary } from '../checks';
+import { NavDictionary } from '../translation-checkers';
 
 export default NavDictionary({
 	startHere: 'Главни',
